Make sure the base url is set when calling authenticate_request
This is a fix for the issue where the base url was not being set when calling authenticate_request.
This caused an error later in the kiota-authentication-oauth-ruby library. The request_info is passed
to the access_token_provider and request_info.uri is used to build the request. This uri is missing the
base url.
https://github.com/microsoft/kiota-abstractions-ruby/blob/7367a156276923eddd22e52335adbba88b1bf71e/lib/microsoft_kiota_abstractions/authentication/base_bearer_token_authentication_provider.rb#L21
Further down the line, the uri is parsed and an error is raised if scheme is not https. Because the base url
is missing, the URI is only parsed as a relative path and thus always fails the scheme check.
https://github.com/microsoft/kiota-authentication-oauth-ruby/blob/69ebeb6705b1d621a9943a15d634e9a33c77d23e/lib/microsoft_kiota_authentication_oauth/oauth_access_token_provider.rb#L54